Decompose `lib.rs` into separate modules
