+++ Cargo.toml
@@ -26,3 +26,6 @@ pyo3 = "0.19.2"
 rand_chacha = "0.3.1"
 rsa = "0.10.0-pre.1"
 tiny-bip39 = "1.0.0"
+
+[dev-dependencies]
+rand = "0.8.5"
